flower support ddd

+++ flower.ddd/src/test/java/com/ly/train/flower/ddd/api/command/CreateOrderCommand.java
@@ -1,3 +1,18 @@
+/**
+ * Copyright © 2019 同程艺龙 (zhihui.li@ly.com)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package com.ly.train.flower.ddd.api.command;
 
 import java.io.Serializable;
@@ -10,22 +25,22 @@ public class CreateOrderCommand implements Serializable {
 
   private static final long serialVersionUID = 1L;
   @TargetAggregateIdentifier
-  private Long id;
+  private Integer id;
   private String name;
 
   /**
    * 
    */
-  public CreateOrderCommand(Long id, String name) {
+  public CreateOrderCommand(Integer id, String name) {
     this.id = id;
     this.name = name;
   }
 
-  public Long getId() {
+  public Integer getId() {
     return id;
   }
 
-  public void setId(Long id) {
+  public void setId(Integer id) {
     this.id = id;
   }
